Even more octave fixes
It seems the gnuplot backend allows saving data under octave
for linux. Under macosc gnuplot the saved files are all black
but even that is better than the fltk backend that crashes
octave. For Linux the issue with fltk might just be that the
plots contain NaNs which ghostscript does not seem to like,
gnuplot ignores NaNs instead...

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>